backend/requirements.txt
azure-functions==1.17.0
azure-messaging-webpubsubservice>=1.1.0